Refactor (APIMF-3491): extracted union parsing and necessary pure functions in related classes/traits
